Feature/hpc update

+++ utils/run_nco_combine_ant_bio.sh
@@ -3,9 +3,17 @@
 #CMAQv5.2.1 CB6 Emission Species w/biogenics
 declare -a species=( "AACD" "ACET" "ACROLEIN" "ALD2" "ALD2_PRIMARY" "ALDX" "APIN" "BENZ" "BUTADIENE13" "CH4" "CH4_INV" "CL2" "CO" "CO2_INV" "ETH" "ETHA" "ETHY" "ETOH" "FACD" "FORM" "FORM_PRIMARY" "HCL" "HONO" "IOLE" "ISOP" "KET" "MEOH" "NAPH" "NH3" "NH3_FERT" "NO" "NO2" "OLE" "PAL" "PAR" "PCA" "PCL" "PEC" "PFE" "PH2O" "PK" "PMC" "PMG" "PMN" "PMOTHR" "PNA" "PNCOM" "PNH4" "PNO3" "POC" "PRPA" "PSI" "PSO4" "PTI" "SESQ" "SO2" "SOAALK" "SULF" "TERP" "TOL" "UNK" "UNR" "VOC_INV" "XYLMN" )
 
+
+# expect arguments 
+if [ $# -lt 2 ]; then
+  echo "Expected two arguments (input and output paths)"
+  echo "Got: $@"
+  exit 2
+fi
+
 #Set input and output files
-input=./output/NEXUS.merged_nei2016v1_meganv21_pre
-output=./output/NEXUS.merged_nei2016v1
+input=$1 
+output=$2
 rm -f $output
 
 #Process emissions
